+++ ChangeLog.md
@@ -23,6 +23,7 @@
 
 - [`KT-7492`](https://youtrack.jetbrains.com/issue/KT-7492) Support "Create abstract function/property" inside an abstract class
 - [`KT-10668`](https://youtrack.jetbrains.com/issue/KT-10668) Support "Create member/extension" corresponding to the extension receiver of enclosing function
+- Show versions in inspection about different version of Kotlin plugin in Maven and IDE plugin
 
 #### Refactorings